fix(ui): resolve ESLint on the drill-down list (low-contrast + react-refresh)
CI's ESLint failed on ExpandedItemsList.jsx: the sort-arrow span used a bare
light-mode `text-gray-400`, which trips the no-low-contrast-text WCAG rule, and the
named exports (rowType / sortItems / itemsToCsv) tripped
react-refresh/only-export-components.

Move the pure helpers into ExpandedItemsList.helpers.js (matching the existing
RecentChangesSection.helpers convention) so the component file exports only a
component, and let the sort arrow inherit the header's compliant colour. No
behaviour change; the list tests import the helpers from the new module.
